% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_modified.R
\name{run_clustering}
\alias{run_clustering}
\title{Run Clustering}
\usage{
run_clustering(
  input_data_list,
  varselect = list(all = "all"),
  methods = c("kmeans", "hierarchical", "dbscan", "vae"),
  params = list(vae = list(latent_dim = 2, hidden_dim = 128, epochs = 10, batch_size =
    32, n_clusters = 5, seed = 42), dbscan = list(eps = c("0.5", "0.75"), minPts = c("2",
    "3")), hierarchical = list(cut_quantile = c(".5"), cutpoint = c("3"), kcut = c("5")),
    kmeans = list(n_clusters = c("2", "3")))
)
}
\arguments{
\item{input_data_list}{A list of dataframes to apply clustering to.}

\item{varselect}{A named list of variable sets to subset data by before clustering. Default is 'all' variables.}

\item{methods}{A character vector of clustering methods to apply. Options are "kmeans", "hierarchical", "dbscan", and "vae".}

\item{params}{A list of parameters specific to each clustering method. Includes settings for each method.}
}
\value{
A list containing clustering results for each method and variable set.
}
\description{
Applies clustering methods to a list of dataframes using specified variable sets and parameters.
}
