% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_modified.R
\name{standardize_cluster_output}
\alias{standardize_cluster_output}
\title{Standardize Cluster Output}
\usage{
standardize_cluster_output(results_list)
}
\arguments{
\item{results_list}{A list containing clustering results.}
}
\value{
A list with two dataframes: `cluster_membership` and `cluster_parameters`.
}
\description{
Standardizes and integrates clustering results for easier analysis.
}
