% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{supervised_rulefit}
\alias{supervised_rulefit}
\title{Fit RuleFit Model and Generate Output Table}
\usage{
supervised_rulefit(
  input_data_list,
  varselect = list(all = "all"),
  dependent_variable,
  family = "gaussian",
  max_depth = 3
)
}
\arguments{
\item{input_data_list}{A list of data frames to apply the RuleFit model to.}

\item{varselect}{A named list of variable sets to subset data by before modeling. Default is 'all' variables.}

\item{dependent_variable}{The dependent variable for the RuleFit model (required).}

\item{family}{The family of the model. Default is "gaussian".}

\item{max_depth}{The maximum depth of the tree used to generate in the RuleFit model. Default is 10.}
}
\value{
A list of data frames containing RuleFit model coefficients for each dataset and variable set.
}
\description{
This function fits a RuleFit model to an input dataset and generates a table of model coefficients.
}
